requirements.txt
anthropic>=0.25.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.127.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
